up to 2.0 step 12
- store token in Connection
- http connection use rpc for most request(import and export yse the old way)
- remove authenticate/invalidate methods
