+++ tests/integration_tests/test_http_connection.py
@@ -89,7 +89,7 @@ class TestHttpConnection(TestCase):
         self.assertIsNotNone(res.result)
         self.assertEqual(res.status, "OK")
         self.assertEqual(res.result["id"], f"article:⟨{uid}⟩")
-        res = connection.select("article", uid)
+        res = connection.select("article", f"⟨{uid}⟩")
         self.assertTrue(res.result != [], f"{uid} - {res}")
         self.assertEqual(res.status, "OK")
         self.assertEqual(res.result[0]["id"], f"article:⟨{uid}⟩", res)
@@ -103,7 +103,7 @@ class TestHttpConnection(TestCase):
         self.assertTrue(res.result != [])
         self.assertEqual(res.status, "OK")
         self.assertEqual(res.result["id"], f"article:⟨{uid}⟩", res)
-        res = connection.select("article", uid)
+        res = connection.select("article", f"⟨{uid}⟩")
         self.assertTrue(res.result != [])
         self.assertEqual(res.status, "OK")
         self.assertEqual(res.result[0]["id"], f"article:⟨{uid}⟩", res)
@@ -124,12 +124,11 @@ class TestHttpConnection(TestCase):
         connection.use("test", "test")
         uid = get_uuid()
         connection.create("article", {"author": uid, "title": uid, "text": uid}, record_id=uid)
-        res = connection.update("article", {"author": "inserted"}, record_id=uid)
-        self.assertTrue(len(res.result) == 1)
+        res = connection.update(f"article:⟨{uid}⟩", {"author": "inserted"})
         self.assertEqual(res.status, "OK")
-        self.assertEqual(res.result[0]["id"], f"article:⟨{uid}⟩", res)
-        self.assertEqual(res.result[0]["author"], "inserted", res)
-        res = connection.select("article", uid)
+        self.assertEqual(res.result["id"], f"article:⟨{uid}⟩", res)
+        self.assertEqual(res.result["author"], "inserted", res)
+        res = connection.select("article", f"`{uid}`")
         self.assertEqual(res.status, "OK")
         self.assertEqual(len(res.result[0]), 2, res)
         self.assertEqual(res.result[0]["id"], f"article:⟨{uid}⟩", res)
@@ -163,12 +162,11 @@ class TestHttpConnection(TestCase):
         connection.use("test", "test")
         uid = get_uuid()
         connection.create("article", {"author": uid, "title": uid, "text": uid}, record_id=uid)
-        res = connection.merge("article", {"new_field": "merge"}, record_id=uid)
-        self.assertTrue(len(res.result) == 1)
+        res = connection.merge("article", {"new_field": "merge"}, record_id=f"⟨{uid}⟩")
         self.assertEqual(res.status, "OK")
-        self.assertEqual(res.result[0]["id"], f"article:⟨{uid}⟩", res)
-        self.assertEqual(res.result[0]["new_field"], "merge", res)
-        res = connection.select("article", uid)
+        self.assertEqual(res.result["id"], f"article:⟨{uid}⟩", res)
+        self.assertEqual(res.result["new_field"], "merge", res)
+        res = connection.select(f"article:⟨{uid}⟩")
         self.assertEqual(res.status, "OK")
         self.assertEqual(res.result[0]["id"], f"article:⟨{uid}⟩", res)
         self.assertEqual(res.result[0]["new_field"], "merge", res)
@@ -207,10 +205,9 @@ class TestHttpConnection(TestCase):
         connection.use("test", "test")
         uid = get_uuid()
         connection.create("article", {"author": uid, "title": uid, "text": uid}, record_id=uid)
-        res = connection.delete("article", record_id=uid)
-        self.assertTrue(len(res.result) == 1)
+        res = connection.delete("article", record_id=f"⟨{uid}⟩")
         self.assertEqual(res.status, "OK")
-        self.assertEqual(res.result[0]["id"], f"article:⟨{uid}⟩", res)
+        self.assertEqual(res.result["id"], f"article:⟨{uid}⟩", res)
         res = connection.select("article", uid)
         self.assertEqual(res.status, "OK")
         self.assertTrue(len(res.result) == 0)
@@ -343,7 +340,7 @@ class TestHttpConnection(TestCase):
             self.assertFalse(res.is_error(), res)
             self.assertIsNotNone(res.query)
             self.assertEqual({'db': 'test', 'http_origin': None, 'ip': '127.0.0.1', 'ns': 'test',
-                              'scope': None, 'session_id': None}, res.result)
+                              'access': None, 'session_id': None}, res.result)
 
     def test_remove_table_with_record(self):
         surreal = Surreal(URL, credentials=('root', 'root'), use_http=True)
@@ -386,18 +383,18 @@ class TestHttpConnection(TestCase):
             self.assertFalse(res.is_error())
             self.assertEqual(res.result, [{'id': f"article:{uid}", 'field': 'old'}])
 
-    def test_update_creates_if_not_exists(self):
+    def test_update_not_creates_if_not_exists(self):
         surreal = Surreal(URL, credentials=('root', 'root'), use_http=True)
         with surreal.connect() as connection:
             connection.use("test", "test")
             uid = get_random_series(13)
-            res = connection.update(f"article:{uid}", {'field': 'old'})
+            res = connection.update(f"article:`{uid}`", {'field': 'old'})
             self.assertFalse(res.is_error())
-            res = connection.select(f"article:{uid}")
+            res = connection.select(f"article:`{uid}`")
             self.assertFalse(res.is_error())
-            self.assertEqual(res.result, [{'id': f"article:{uid}", 'field': 'old'}])
+            self.assertEqual(res.result, [])
 
-    def test_merge_creates_if_not_exists(self):
+    def test_merge_not_creates_if_not_exists(self):
         surreal = Surreal(URL, credentials=('root', 'root'), use_http=True)
         with surreal.connect() as connection:
             connection.use("test", "test")
@@ -406,7 +403,7 @@ class TestHttpConnection(TestCase):
             self.assertFalse(res.is_error())
             res = connection.select(f"article:{uid}")
             self.assertFalse(res.is_error())
-            self.assertEqual(res.result, [{'id': f"article:{uid}", 'field': 'old'}])
+            self.assertEqual(res.result, [])
 
     def test_delete_unexisting(self):
         surreal = Surreal(URL, credentials=('root', 'root'), use_http=True)
@@ -414,7 +411,7 @@ class TestHttpConnection(TestCase):
             connection.use("test", "test")
             uid = get_random_series(14)
             res = connection.delete(f"article:{uid}")
-            self.assertEqual(res.result, [None])
+            self.assertEqual(res.result, None)
             self.assertFalse(res.is_error())
             res = connection.delete(uid)
             self.assertFalse(res.is_error())
@@ -443,8 +440,8 @@ class TestHttpConnection(TestCase):
             connection.create("article", {"id": uid, "author": uid, "title": uid, "text": uid})
             res = connection.patch("article", [{"op": "replace", "path": "/active", "value": True}], uid)
             self.assertFalse(res.is_error(), res)
-            self.assertEqual(res.result, [{"id": f"article:{uid}", "author": uid, "title": uid, "text": uid,
-                                           "active": True}])
+            self.assertEqual(res.result, {"id": f"article:{uid}", "author": uid, "title": uid, "text": uid,
+                                          "active": True})
             res = connection.select(f"article:{uid}")
             self.assertFalse(res.is_error(), res)
             self.assertEqual(res.result[0]['author'], uid)
@@ -458,7 +455,7 @@ class TestHttpConnection(TestCase):
             connection.create("article", {"id": uid, "author": uid, "title": uid, "text": uid, "active": False})
             res = connection.patch("article", [{"op": "replace", "path": "/active", "value": True}], uid, True)
             self.assertFalse(res.is_error(), res)
-            self.assertEqual(res.result, [[{'op': 'replace', 'path': '/active', 'value': True}]])
+            self.assertEqual(res.result, [{'op': 'replace', 'path': '/active', 'value': True}])
             res = connection.select(f"article:{uid}")
             self.assertFalse(res.is_error(), res)
             self.assertEqual(res.result[0]['author'], uid)
